+++ pkg/controller/customresource/customresource_test.go
@@ -55,28 +55,28 @@ func TestReconciliationShouldBeSkipped(t *testing.T) {
 
 	t.Run("Empty annotations", func(t *testing.T) {
 		for _, resourceType := range newResourceTypes() {
-			assert.False(t, ReconciliationShouldBeSkipped(resourceType))
+			assert.False(t, ReconciliationShouldBeSkipped(resourceType, ""))
 		}
 	})
 
 	t.Run("Other resource types", func(t *testing.T) {
 		for _, resourceType := range newResourceTypes() {
 			resourceType.SetAnnotations(map[string]string{"foo": "bar"})
-			assert.False(t, ReconciliationShouldBeSkipped(resourceType))
+			assert.False(t, ReconciliationShouldBeSkipped(resourceType, ""))
 		}
 	})
 
 	t.Run("Annotation present, reconciliation should not be skipped", func(t *testing.T) {
 		for _, resourceType := range newResourceTypes() {
 			resourceType.SetAnnotations(map[string]string{ReconciliationPolicyAnnotation: "foobar"})
-			assert.False(t, ReconciliationShouldBeSkipped(resourceType))
+			assert.False(t, ReconciliationShouldBeSkipped(resourceType, ""))
 		}
 	})
 
 	t.Run("Annotation present, reconciliation should be skipped", func(t *testing.T) {
 		for _, resourceType := range newResourceTypes() {
 			resourceType.SetAnnotations(map[string]string{ReconciliationPolicyAnnotation: ReconciliationPolicySkip})
-			assert.True(t, ReconciliationShouldBeSkipped(resourceType))
+			assert.True(t, ReconciliationShouldBeSkipped(resourceType, ""))
 		}
 	})
 }